Bugfix for Chrome 53, support for Chrome <52
Previous to Chrome 53, SVG origin was set at the top of the window, vs the way in which Firefox sets SVG origin: the top of the parent element. I've added a conditional to check for the current version of Chrome, and have offset the SVG clipPath points based on that version number.